Fix comments by reviewer

+++ src/main/java/cz/muni/fi/pv168/project/data/TestDataGenerator.java
@@ -2,6 +2,7 @@ package cz.muni.fi.pv168.project.data;
 
 import cz.muni.fi.pv168.project.model.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
@@ -37,8 +38,8 @@ public class TestDataGenerator {
 
     public Ride createTestRide() {
         Currency currency = selectRandom(Arrays.stream(Currency.values()).toList());
-        float distance = random.nextFloat(MAX_DISTANCE+1);
-        float price = random.nextFloat(MAX_PRICE+1);
+        BigDecimal distance = BigDecimal.valueOf(random.nextFloat(MAX_DISTANCE+1));
+        BigDecimal price = BigDecimal.valueOf(random.nextFloat(MAX_PRICE+1));
         int passengersCount = random.nextInt(1, MAX_PASSENGERS+1);
         LocalDateTime dateTime = selectRandomLocalDateTime(MIN_DATE, MAX_DATE);
         Category category = selectRandom(CATEGORIES);
